Refactor: Extract shared name utilities to separate module
Move RENAME_MAP, check_rename, name_to_underscore_name, and name_to_ident
from lib.rs and main.rs into a new name_utils.rs module to eliminate
code duplication and prevent drift between the two.
